backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
lxml==4.9.3
beautifulsoup4==4.12.2
pdfminer.six==20221105
reportlab==4.0.6
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
aiofiles==23.2.0
pydantic==2.5.0
python-magic==0.4.27
ebooklib==0.18